Day 0909 : 쿠폰생성로직

+++ src/main/java/com/ubic/shop/service/ShopListService.java
@@ -30,7 +30,7 @@ public class ShopListService {
 
     /** 장바구니 */
     @Transactional
-    public Long shopList(Long userId, Long productId, int count) {
+    public Long shopList(Long userId, Long productId, long count) {
         //엔티티 조회
         User user = userRepository.findById(userId).get();
         Product product = productService.findById(productId);
